[MODIFY] pet 관리하는 api 구조 변경 및 swagger 추가
